+++ Kernel/System/Time.pm
@@ -443,7 +443,7 @@ sub WorkingTime {
             );
             my $Zone = $ConfigObject->Get( "TimeZone::Calendar" . $Param{Calendar} );
             if ($Zone) {
-                $Zone = $Zone * 3600;    # 60 * 60
+                $Zone *= 3600;
                 $Param{StartTime} += $Zone;
                 $Param{StopTime}  += $Zone;
             }
@@ -463,21 +463,22 @@ sub WorkingTime {
     my $Counted = 0;
     my ( $ASec, $AMin, $AHour, $ADay, $AMonth, $AYear, $AWDay )
         = localtime $Param{StartTime};    ## no critic
-    $AYear  = $AYear + 1900;
-    $AMonth = $AMonth + 1;
+    $AYear  += 1900;
+    $AMonth += 1;
     my $ADate = "$AYear-$AMonth-$ADay";
     my ( $BSec, $BMin, $BHour, $BDay, $BMonth, $BYear, $BWDay )
         = localtime $Param{StopTime};     ## no critic
-    $BYear  = $BYear + 1900;
-    $BMonth = $BMonth + 1;
+    $BYear  += 1900;
+    $BMonth += 1;
     my $BDate = "$BYear-$BMonth-$BDay";
 
     while ( $Param{StartTime} < $Param{StopTime} ) {
         my ( $Sec, $Min, $Hour, $Day, $Month, $Year, $WDay )
             = localtime $Param{StartTime};    ## no critic
-        $Year  = $Year + 1900;
-        $Month = $Month + 1;
+        $Year  += 1900;
+        $Month +=  1;
         my $CDate = "$Year-$Month-$Day";
+        my $CTime00 = $Param{StartTime} - ( ( $Hour * 60 + $Min ) * 60 + $Sec );    # 00:00:00
 
         # count nothing because of vacation
         if (
@@ -490,30 +491,30 @@ sub WorkingTime {
         }
         else {
             if ( $TimeWorkingHours->{ $LDay{$WDay} } ) {
-                for ( @{ $TimeWorkingHours->{ $LDay{$WDay} } } ) {
+                for my $WorkingHour ( @{ $TimeWorkingHours->{ $LDay{$WDay} } } ) {
 
-                    # count minutes on same date and same hour of start/end date
-                    # within service hour => start counting and finish immediatly
-                    if ( $ADate eq $BDate && $AHour == $BHour && $AHour == $_ ) {
-                        return ( ( $BMin - $AMin ) * 60 );
+                    # same date and same hour of start/end date within service hour
+                    # => start counting and finish immediatly
+                    if ( $ADate eq $BDate && $AHour == $BHour && $AHour == $WorkingHour ) {
+                        return $Param{StopTime} - $Param{StartTime};
                     }
 
                     # do nothing because we are on start day and not yet within service hour
-                    elsif ( $CDate eq $ADate && $_ < $AHour ) {
+                    elsif ( $CDate eq $ADate && $WorkingHour < $AHour ) {
                     }
 
-                    # count minutes because we are on start day and within start hour
-                    elsif ( $CDate eq $ADate && $AHour == $_ ) {
-                        $Counted = $Counted + ( 60 - $AMin ) * 60;
+                    # we are on start day and within start hour => count to end of this hour
+                    elsif ( $CDate eq $ADate && $AHour == $WorkingHour ) {
+                        $Counted += ( $CTime00 + ( $WorkingHour + 1 ) * 60 * 60 ) - $Param{StartTime};
                     }
 
                     # do nothing because we are on end day but greater than service hour
-                    elsif ( $CDate eq $BDate && $BHour < $_ ) {
+                    elsif ( $CDate eq $BDate && $BHour < $WorkingHour ) {
                     }
 
-                    # count minutes because we are on end day and within end hour
-                    elsif ( $CDate eq $BDate && $BHour == $_ ) {
-                        $Counted = $Counted + $BMin * 60;
+                    # we are on end day and within end hour => count from start of this hour
+                    elsif ( $CDate eq $BDate && $BHour == $WorkingHour ) {
+                        $Counted += $Param{StopTime} - ( $CTime00 + $WorkingHour * 60 * 60 );
                     }
 
                     # count full hour because we are in service hour that is greater than
@@ -526,7 +527,14 @@ sub WorkingTime {
         }
 
         # reduce time => go to next day 00:00:00
-        $Param{StartTime} = $Param{StartTime} + 60 * 60 * ( 24 - $Hour ) - 60 * $Min - $Sec;
+        $Param{StartTime} = $Self->Date2SystemTime(
+            Year   => $Year,
+            Month  => $Month,
+            Day    => $Day,
+            Hour   => 23,
+            Minute => 59,
+            Second => 59,
+        ) + 1;
     }
     return $Counted;
 }
@@ -534,10 +542,10 @@ sub WorkingTime {
 =item DestinationTime()
 
 get the destination time based on the current calendar working time (fallback: default
-system working time) configuragtion.
+system working time) configuration.
 
 The algorithm roughly works as follows:
-    - Check if the start time is acutally in the configured working time.
+    - Check if the start time is actually in the configured working time.
         - If not, set it to the next working time second. Example: start time is
             on a weekend, start time would be set to 8:00 on the following Monday.
     - Then the diff time (in seconds) is added to the start time incrementally, only considering
@@ -545,7 +553,7 @@ The algorithm roughly works as follows:
         they would be spread over the configured working hours. If we have 8-20, 24 hours would be
         spread over 2 days (13/11 hours).
 
-NOTE: Currently, the implementation stops silently after 100 iterations, making it impossible to
+NOTE: Currently, the implementation stops silently after 600 iterations, making it impossible to
     specify longer escalation times, for example.
 
     my $DestinationTime = $TimeObject->DestinationTime(
@@ -594,14 +602,12 @@ sub DestinationTime {
                 "TimeVacationDaysOneTime::Calendar" . $Param{Calendar}
             );
             $Zone = $ConfigObject->Get( "TimeZone::Calendar" . $Param{Calendar} );
-            $Zone = $Zone * 3600;                                                    # 60 * 60
-            $Param{StartTime} = $Param{StartTime} + $Zone;
+            $Zone *= 3600;
+            $Param{StartTime} += $Zone;
         }
     }
     my $DestinationTime = $Param{StartTime};
     my $CTime           = $Param{StartTime};
-    my $FirstTurn       = 1;
-    $Param{Time}++;
 
     my %LDay = (
         1 => 'Mon',
@@ -618,11 +624,12 @@ sub DestinationTime {
     LOOP:
     while ( $Param{Time} > 1 ) {
         $LoopCounter++;
-        last LOOP if $LoopCounter > 100;
+        last LOOP if $LoopCounter > 600;
 
         my ( $Second, $Minute, $Hour, $Day, $Month, $Year, $WDay ) = localtime $CTime;  ## no critic
-        $Year  = $Year + 1900;
-        $Month = $Month + 1;
+        $Year  += 1900;
+        $Month += 1;
+        my $CTime00 = $CTime - ( ( $Hour * 60 + $Minute ) * 60 + $Second );             # 00:00:00
 
         # Skip vacation days, or days without working hours, do not count.
         if (
@@ -631,19 +638,15 @@ sub DestinationTime {
             || !$TimeWorkingHours->{ $LDay{$WDay} }
             )
         {
-
-            if ($FirstTurn) {
-                $DestinationTime = $Self->Date2SystemTime(
-                    Year   => $Year,
-                    Month  => $Month,
-                    Day    => $Day,
-                    Hour   => 0,
-                    Minute => 0,
-                    Second => 0,
-                );
-            }
-            $DestinationTime = $DestinationTime + 60 * 60 * 24;
-            $FirstTurn       = 0;
+            # Set destination time to next day, 00:00:00
+            $DestinationTime = $Self->Date2SystemTime(
+                Year   => $Year,
+                Month  => $Month,
+                Day    => $Day,
+                Hour   => 23,
+                Minute => 59,
+                Second => 59,
+            ) + 1;
         }
 
         # Regular day with working hours
@@ -654,48 +657,25 @@ sub DestinationTime {
                 # Check if we have a working hour
                 if ( grep { $H == $_ } @{ $TimeWorkingHours->{ $LDay{$WDay} } } ) {
                     if ( $Param{Time} > 60 * 60 ) {
-                        if ( $Minute != 0 && $FirstTurn ) {
-                            my $Max = 60 - $Minute;
-                            $Param{Time} = $Param{Time} - ( $Max * 60 );
-                            $DestinationTime = $DestinationTime + ( $Max * 60 );
-                            $FirstTurn = 0;
-                        }
-                        else {
-                            $Param{Time} = $Param{Time} - ( 60 * 60 );
-                            $DestinationTime = $DestinationTime + ( 60 * 60 );
-                            $FirstTurn = 0;
-                        }
-                    }
-                    elsif ( $Param{Time} > 1 * 60 ) {
-                        for my $M ( 0 .. 59 ) {
-                            if ( $Param{Time} > 1 ) {
-                                $Param{Time}     = $Param{Time} - 60;
-                                $DestinationTime = $DestinationTime + 60;
-                                $FirstTurn       = 0;
-                            }
-                        }
+                        my $RestOfHour = 3600 - ( $Minute * 60 + $Second );
+                        $DestinationTime += $RestOfHour;
+                        $Param{Time} -= $RestOfHour;
                     }
                     else {
-                        last HOUR;
+                        $DestinationTime += $Param{Time};
+                        last LOOP;
                     }
                 }
 
                 # Not a working hour
                 else {
-                    if ($FirstTurn) {
-                        $DestinationTime = $Self->Date2SystemTime(
-                            Year   => $Year,
-                            Month  => $Month,
-                            Day    => $Day,
-                            Hour   => $H,
-                            Minute => 0,
-                            Second => 0,
-                        );
-                    }
-                    if ( $Param{Time} > 59 ) {
-                        $DestinationTime = $DestinationTime + ( 60 * 60 );
-                    }
+                    my $RestOfHour = 3600 - ( $Minute * 60 + $Second );
+                    $DestinationTime += $RestOfHour;
                 }
+
+                # Here we are always aligned at an hour boundary
+                $Minute = 0;
+                $Second = 0;
             }
         }
 
@@ -704,46 +684,20 @@ sub DestinationTime {
             Year   => $Year,
             Month  => $Month,
             Day    => $Day,
-            Hour   => 0,
-            Minute => 0,
-            Second => 0,
-        ) + ( 60 * 60 * 24 );
+            Hour   => 23,
+            Minute => 59,
+            Second => 59,
+        ) + 1;
 
-        # Protect local time zone problems on your machine
-        # (e. g. summertime -> wintertime) and not getting over to the next day.
-        if ( $NewCTime == $CTime ) {
-            $CTime = $CTime + ( 60 * 60 * 24 );
-
-            # reduce destination time diff between today and tomorrow
-            my ( $NextSecond, $NextMinute, $NextHour, $NextDay, $NextMonth, $NextYear )
-                = localtime $CTime;    ## no critic
-            $NextYear  = $NextYear + 1900;
-            $NextMonth = $NextMonth + 1;
-
-            my $Diff = (
-                $Self->Date2SystemTime(
-                    Year   => $NextYear,
-                    Month  => $NextMonth,
-                    Day    => $NextDay,
-                    Hour   => 0,
-                    Minute => 0,
-                    Second => 0,
-                    ) - $Self->Date2SystemTime(
-                    Year   => $Year,
-                    Month  => $Month,
-                    Day    => $Day,
-                    Hour   => 0,
-                    Minute => 0,
-                    Second => 0,
-                    )
-            ) - ( 60 * 60 * 24 );
-            $DestinationTime = $DestinationTime - $Diff;
+        # Compensate for switching to/from daylight saving time
+        # (day is shorter or longer than 24h)
+        if ( $NewCTime != $CTime00 + 24 * 60 * 60 ) {
+            my $Diff = $NewCTime - $CTime00 - 24 * 60 * 60;
+            $DestinationTime += $Diff;
         }
 
         # Set next loop time to 00:00:00 of next day.
-        else {
-            $CTime = $NewCTime;
-        }
+        $CTime = $NewCTime;
     }
 
     # return destination time - e. g. with diff of calendar time zone